fix: crash on web
